user profile changed & added alert
